Serverless hapi (#118)
* adding serverless handler

* moving directory back and updating index handler

* adding better now.json config

* logging url in config

* updating alias
