+++ api/.gitignore
@@ -5,3 +5,4 @@
 node_modules
 dist
 docs
+!package-lock.json
