+++ api/src/util/hash.ts
