+++ client/utils/urls.ts
@@ -1,7 +1,7 @@
 import { __IS_DEV__ } from "./dev";
 
 const DEV_URL = "http://localhost:5000/api";
-const PROD_URL = "https://beta.downwrite.us/api";
+const PROD_URL = "https://next.downwrite.us/api";
 
 const URL: string = __IS_DEV__ ? DEV_URL : PROD_URL;
 
@@ -16,3 +16,5 @@ export const PASSWORD_ENDPOINT: string = `${URL}/password`;
 export const SETTINGS_ENDPOINT: string = `${URL}/users/settings`;
 
 export const AUTH_ENDPOINT: string = `${URL}/users/authenticate`;
+
+export const prefixURL = (p: string, url: string): string => p + url;
