+++ now.json
@@ -2,6 +2,7 @@
   "version": 2,
   "public": true,
   "name": "downwrite",
+  "alias": ["next.downwrite.us"],
   "env": {
     "PUPPETEER_SKIP_CHROMIUM_DOWNLOAD": "true",
     "PORT": "4411",
@@ -12,10 +13,10 @@
   },
   "builds": [
     { "src": "client/next.config.js", "use": "@now/next" },
-    { "src": "api/server.ts", "use": "@now/node@canary" }
+    { "src": "api/src/**/*.ts", "use": "@now/node@canary" }
   ],
   "routes": [
-    { "src": "/api/(.*)", "dest": "/api/$1" },
+    { "src": "/api/(.*)", "dest": "/api/src/index.ts" },
     { "src": "/(.*)", "dest": "/client/$1" }
   ]
 }
